# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse____": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {},
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_any_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmojiType": {
				"enum": [
					"like",
					"laugh",
					"cry",
					"angry"
				],
				"type": "string"
			},
			"EmojiInput": {
				"properties": {
					"emoji": {
						"$ref": "#/components/schemas/EmojiType"
					}
				},
				"required": [
					"emoji"
				],
				"type": "object"
			},
			"CommentInput": {
				"properties": {
					"parentId": {
						"type": "string"
					},
					"authorId": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"authorId",
					"postId",
					"content"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "UNNERA API",
		"version": "1.0.0",
		"description": "API Docs for UNERRA",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "utkarsh yadav"
		}
	},
	"paths": {
...
		"/post/create": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"description": "Create a new post",
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"content": {
										"type": "string"
									},
									"tags": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"markdown": {
										"type": "string"
									},
									"videoUrl": {
										"type": "string"
									},
									"imageUrl": {
										"type": "string"
									},
									"embedUrl": {
										"type": "string"
									},
									"isPublished": {
										"type": "string"
									},
									"publishedAt": {
										"type": "string"
									},
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"title",
									"content"
								]
							}
						}
					}
				}
			}
		},
		"/post": {
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/post/{id}": {
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse____"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"content": {
										"type": "string"
									},
									"tags": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"markdown": {
										"type": "string"
									},
									"videoUrl": {
										"type": "string"
									},
									"imageUrl": {
										"type": "string"
									},
									"embedUrl": {
										"type": "string"
									},
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								}
							}
						}
					}
				}
			}
		},
		"/post/{id}/likes": {
			"get": {
				"operationId": "GetPostLikes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/post/{id}/like": {
			"post": {
				"operationId": "ToggleLike",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmojiInput"
							}
						}
					}
				}
			}
		},
		"/post/{id}/comment": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CommentInput"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}